+++ app/renderer/reducers/Session.js
@@ -25,6 +25,9 @@ const INITIAL_STATE = {
       dataCenter: 'US',
     },
     headspin: {},
+    browserstack: {},
+    advanced: {},
+    bitbar: {},
   },
   attachSessId: null,
 
@@ -173,7 +176,7 @@ export default function session (state = INITIAL_STATE, action) {
     case SET_SERVER:
       return {
         ...state,
-        // Only set remote, sauce, headspin, and testobject;
+        // Only set remote and cloud providers;
         // 'local' comes from electron-settings
         server: {
           ...state.server,
@@ -181,6 +184,8 @@ export default function session (state = INITIAL_STATE, action) {
           sauce: action.server.sauce || {},
           testobject: action.server.testobject || {},
           headspin: action.server.headspin || {},
+          browserstack: action.server.browserstack || {},
+          bitbar: action.server.bitbar || {},
         },
         serverType: action.serverType || ServerTypes.local,
       };