+++ app/shared/cloud-providers/config.js
@@ -1,0 +1,17 @@
+export default {
+  saucelabs: {
+    label: 'Sauce Labs',
+  },
+  testobject: {
+    label: 'TestObject'
+  },
+  browserstack: {
+    label: 'BrowserStack',
+  },
+  headspin: {
+    label: 'HeadSpin',
+  },
+  bitbar: {
+    label: 'Bitbar'
+  }
+};