+++ test/e2e/main-e2e.test.js
@@ -49,7 +49,7 @@ describe('application launch', function () {
 
   it('starts the server and opens a new session window', async function () {
     // Start the server
-    await client.waitForExist(main.startServerButton);
+    await client.waitForExist(main.startServerButton, 30000);
     await main.startServer();
 
     // Wait for the server monitor container to be present
@@ -71,10 +71,11 @@ describe('application launch', function () {
     }
 
     const resourcesWDAPath = path.join(__dirname, '..', '..', 'release', 'mac', 'Appium.app', 'Contents', 'Resources', 
-      'app', 'node_modules', 'appium-xcuitest-driver', 'WebDriverAgent');
+      'app', 'node_modules', 'appium', 'node_modules', 'appium-xcuitest-driver', 'WebDriverAgent');
 
-    const localWDAPath = path.join(__dirname, '..', '..', 'node_modules', 'appium-xcuitest-driver', 'WebDriverAgent');
+    await fs.exists(path.join(resourcesWDAPath, 'PrivateHeaders')).should.eventually.be.true;
 
+    const localWDAPath = path.join(__dirname, '..', '..', 'node_modules', 'appium', 'node_modules', 'appium-xcuitest-driver', 'WebDriverAgent');
     const res = await dirCompare.compare(resourcesWDAPath, localWDAPath);
     res.distinct.should.equal(0);
   });